extract test helpers to separate file
